Move issue #581's focus-poll diagnostic into E2EFixture
Pure refactor, no behavior change: SetupActivityPermissionDialogE2ETest's
dumpsys-based window-focus poll moves into E2EFixture.waitForWindowFocus()
so PartialAccessPhotoPickerE2ETest (which drives the same permission
dialog for its own button) can reuse it instead of duplicating the
dumpsys parsing and logging.
